Updating branch 646_update_docs with latest master

+++ utilities/build/run_build.sh
@@ -12,10 +12,13 @@
 #                   root for the build
 #
 #  Optional:
-#     BRANCH      - This is an optional parameter that allows you to configure
+#     BRANCH_NAME   - This is an optional parameter that allows you to configure
 #                   the branch to build
-#     IZPACK_PATH - Path to where IzPack is installed
-#     XTUMLORG_USER - The username for the upload account
+#     UPLOAD_SPEC - This optional argument, if present, will upload the build to this given location.  example: user@myserver.com:/myfolder
+#                   For xtuml.org this is: 
+#					  n5e22526185966@xtuml.org:/home/n5e22526185966/html/wp-content/uploads
+#     package_only - This optional argument, if present and set to "yes" cause the build script to only package and notify
+#     clean - This optional argument, it defaults to yes.  If yes, a clean build if performed.
 #               
 #  Since it is the starting point for the build chain, it must be manually put 
 #  into place for the build server to run. The variable BUILD_MOUNT holds the 
@@ -23,7 +26,7 @@
 #  
 #  Build Server Requirements:
 #  1) run_build.sh, and init_git_repositories.sh must be present in ${BUILD_ROOT}
-#  2) git, ant, jar, and IzPack must be installed on the build server
+#  2) git, ant, and jar must be installed on the build server
 # 
 #  The build is performed under ${BUILD_ROOT}.  The result of the build:
 #  1) plugins
@@ -80,14 +83,16 @@ function distribute_and_notify {
 	  cat ${DIFF_FILE} >> ${MAIL_TEMP}	  
 	fi
 	
-	# TODO - temporary copy location until we upload it somewhere...
 	SERVER_IP=$(dig +short myip.opendns.com @resolver1.opendns.com)
-	SCP_CMD="scp youruser@${SERVER_IP}:${RESULT_FOLDER}/BridgePoint_${BRANCH}_linux.jar BridgePoint_linux.jar"
+	SCP_CMD="scp youruser@${SERVER_IP}:${RESULT_FOLDER}/*.zip"
+	
+	echo -e " " >> ${MAIL_TEMP}
 	echo -e "You can copy the release via: ${SCP_CMD}" >> ${MAIL_TEMP}
 	echo -e "---------------" >> ${MAIL_TEMP}
-	echo -e "The Linux release can be downloaded at: ${DOWNLOAD_URL}_${BRANCH}_linux.jar" >> ${MAIL_TEMP}
-    echo -e "The Windows release can be downloaded at: ${DOWNLOAD_URL}_${BRANCH}_windows.jar" >> ${MAIL_TEMP}
-	
+	echo -e "The Linux release can be downloaded at: ${DOWNLOAD_URL}_${BRANCH}_linux.zip" >> ${MAIL_TEMP}
+        echo -e "The Windows release can be downloaded at: ${DOWNLOAD_URL}_${BRANCH}_windows.zip" >> ${MAIL_TEMP}
+	echo -e " " >> ${MAIL_TEMP}
+
 	cat ${MAIL_TEMP} | ${MAIL_CMD} ${BUILD_ADMIN}
 	
 	rm -rf ${MAIL_TEMP}
@@ -100,13 +105,12 @@ function distribute_and_notify {
 
 # Verify parameters
 if [ "$#" -lt 2 ]; then
-
-echo "This script requires two parameters.  The other parameters are optional.  See below for usage."
-echo
-echo "run_build.sh BridgePoint_Home_Directory Build_Root Branch IzPack_Home_Directory"
-echo
-echo "See the script header for more detail."
-exit 1
+  echo "This script requires two parameters.  The other parameters are optional.  See below for usage."
+  echo
+  echo "run_build.sh BridgePoint_Home_Directory Build_Root <Branch_Name> <upload location> <package only> <clean>"
+  echo
+  echo "See the script header for more detail."
+  exit 1
 fi 
 
 date 
@@ -129,10 +133,19 @@ if [ "$3" != "" ]; then
   export BRANCH="$3"
 fi
 if [ "$4" != "" ]; then
-  export IZPACK_PATH="$4"
+  export UPLOAD_SPEC="$4"
+else
+  export UPLOAD_SPEC=""
 fi
 if [ "$5" != "" ]; then
-  export XTUMLORG_USER="$5"
+  export package_only="$5"
+else
+  export package_only=""
+fi
+if [ "$6" != "" ]; then
+  export clean="$6"
+else
+  export clean="yes"
 fi
 
 # Make sure github credentials are available in the environment
@@ -205,8 +218,10 @@ export DISTRIBUTION_SERVER=""
 # builds, and this was the format of the tag
 export BUILD_TAG="`date +N%F`"
 
-# assure that we are starting with a clean build folder.
-rm -rf ${BUILD_DIR}
+if [ "${package_only}" != "yes" ]; then
+  # assure that we are starting with a clean build folder.
+  rm -rf ${BUILD_DIR}
+fi
 
 mkdir -p "${BUILD_DIR}"
 mkdir -p "${LOG_DIR}"
@@ -226,41 +241,76 @@ if [ $? -eq 1 ]; then
 else
   rm -f "${BUILD_ROOT}/init_git_repositories.tmp"
 fi
-echo -e "Getting files from github, this could take a while."
-bash "${BUILD_ROOT}/init_git_repositories.sh" ${BRANCH} ${GIT_REPO_ROOT} ${ALLOW_FALLBACK} >> ${BUILD_LOG} 2>&1
-echo -e "Done."
+RETVAL=0
+if [ "${package_only}" != "yes" ]; then
+  echo -e "Getting files from github, this could take a while."
+  bash "${BUILD_ROOT}/init_git_repositories.sh" ${BRANCH} ${GIT_REPO_ROOT} ${ALLOW_FALLBACK} >> ${BUILD_LOG} 2>&1
+  echo -e "Done getting files from github."
 
-# Can do the copy and dos2unix translation in one step.
-tr -d '\r' < ${GIT_BP}/utilities/build/configure_build_process.sh > configure_build_process.sh
-chmod a+x configure_build_process.sh
+  # Can do the copy and dos2unix translation in one step.
+  echo -e "Configuring files for the build process."
+  tr -d '\r' < ${GIT_BP}/utilities/build/configure_build_process.sh > configure_build_process.sh
+  chmod a+x configure_build_process.sh
 
-bash configure_build_process.sh >> ${BUILD_LOG}
+  bash configure_build_process.sh >> ${BUILD_LOG}
+  cd  "${BUILD_DIR}"
+  echo -e "Done configuring files for the build process."
+
+  echo -e "Configuring external dependencies."
+  bash configure_external_dependencies.sh ${GIT_REPO_ROOT} > ${LOG_DIR}/configure_externals.log 2>&1
+  cd  "${BUILD_DIR}"
+  echo -e "Done configuring external dependencies."
+
+  echo -e "Building BridgePoint.  This will take a long time."
+  cd $BUILD_DIR
+  tr -d '\r' < ${GIT_BP}/utilities/build/headless_build.sh > headless_build.sh
+  chmod a+x headless_build.sh
+
+  ./headless_build.sh "${BRANCH}" "${BPHOMEDIR}" "${ECLIPSE_HOME}" "${WORKSPACE}" "${GIT_BP}" "${clean}"
+  RETVAL=$?
+  echo -e "Done building."
+
+  # TODO - we'll re-enable this check when headless_build stops reporting errors
+  echo -e "Packaging BridgePoint into a full eclipse environment."
+  #if [ ! -s ${ERROR_FILE} ]; then
+    #exit 1
+  #fi
+  # TODO - we'll re-enable this check when headless_build stops reporting errors
+  #if [ $RETVAL -eq 0 ]; then
+    #exit 1
+  #fi
+fi
+# This packages the build
 cd  "${BUILD_DIR}"
-
-bash configure_external_dependencies.sh ${GIT_REPO_ROOT} > ${LOG_DIR}/configure_externals.log 2>&1
-cd  "${BUILD_DIR}"
-
 bash create_bp_release.sh  >> ${BUILD_LOG}
+
+# Check for errors, if found report them.  Note that the log file is moved after this script runs,
+# hence the different paths for where we grep and where we report the user to look.
+grep -c -i -w "Error" ${BUILD_LOG}
+error_count=$?
+if [ ${error_count} -ne 1 ]; then
+  echo -e "Errors found in the output log. Check ${BUILD_LOG}." >> ${ERROR_FILE}
+fi
+
+if [ -f $ERROR_FILE ]; then
+  echo -e "Errors found during release creation:\n\n\n"
+  cat $ERROR_FILE
+fi
+
+bp_release_version=`awk -F"\"" '{if (/ersion.*\=.*[0-9]\.[0-9]\.[0-9]/) {print $2; exit;}}' ${GIT_BP}/src/org.xtuml.bp.pkg/plugin.xml`
+bash build_installer_bp.sh ${BRANCH} ${STAGING_AREA} ${RESULT_FOLDER} windows ${bp_release_version} "${UPLOAD_SPEC}" >> ${BUILD_LOG}
 cd  "${BUILD_DIR}"
-
-if [ "${IZPACK_PATH}" = "" ]; then
-  export IZPACK_PATH="/usr/local/IzPack"
-fi  
-
-# TODO - we'll re-enable this check when headless_build stops reporting errors
-#if [ ! -s ${ERROR_FILE} ]; then
-  if [ -e ${IZPACK_PATH}/bin/compile ]; then
-    bp_release_version=`awk -F"\"" '{if (/ersion.*\=.*[0-9]\.[0-9]\.[0-9]/) {print $2; exit;}}' ${GIT_BP}/src/org.xtuml.bp.pkg/plugin.xml`
-    bash build_installer_bp.sh ${BRANCH} ${STAGING_AREA} ${IZPACK_PATH} ${RESULT_FOLDER} windows ${bp_release_version} ${XTUMLORG_USER} >> ${BUILD_LOG}
-    cd  "${BUILD_DIR}"
   
-    bash build_installer_bp.sh ${BRANCH} ${STAGING_AREA} ${IZPACK_PATH} ${RESULT_FOLDER} linux ${bp_release_version} ${XTUMLORG_USER} >> ${BUILD_LOG}
-    cd  "${BUILD_DIR}"
-  fi
-#fi
+bash build_installer_bp.sh ${BRANCH} ${STAGING_AREA} ${RESULT_FOLDER} linux ${bp_release_version} "${UPLOAD_SPEC}" >> ${BUILD_LOG}
 
+# This get called regardless of if we are building or packaging to notify the the build is complete
 if [ -e ${MAIL_CMD} ]; then
+  cd  "${BUILD_DIR}"
   distribute_and_notify $? >> ${BUILD_LOG}
 fi 
+chmod -R g+w ${BUILD_DIR}
+chmod -R g+w ${RESULT_FOLDER}
+chmod -R g+w ${GIT_REPO_ROOT}
 
+date
 echo -e "End of run_build.sh"
